requirements.txt
Pygments==2.9.0
Pillow==8.2.0
